display_e2e_alert Display the alert only when the user opens his first encrypted room.
