+++ vector/src/main/java/im/vector/activity/VectorRoomActivity.java
@@ -31,6 +31,7 @@ import android.content.res.Resources;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Build;
+import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 import android.support.v4.app.FragmentManager;
 import android.os.Bundle;
@@ -1065,7 +1066,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements MatrixMe
             mVectorOngoingConferenceCallView.onActivityResume();
         }
 
-        CommonActivityUtils.displayE2eRoomAlert(this, mRoom);
+        displayE2eRoomAlert();
 
         Log.d(LOG_TAG, "-- Resume the activity");
     }
@@ -3062,6 +3063,32 @@ public class VectorRoomActivity extends MXCActionBarActivity implements MatrixMe
                     launchRoomDetails(VectorRoomDetailsActivity.PEOPLE_TAB_INDEX);
                 }
             });
+        }
+    }
+
+    private static final String E2E_WARNINGS_PREFERENCES = "E2E_WARNINGS_PREFERENCES";
+
+    /**
+     * Display an e2e alert for the first opened room.
+     */
+    private void displayE2eRoomAlert() {
+        SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
+
+        if (!preferences.contains(E2E_WARNINGS_PREFERENCES) && (null != mRoom) && mRoom.isEncrypted()) {
+            SharedPreferences.Editor editor = preferences.edit();
+            editor.putBoolean(E2E_WARNINGS_PREFERENCES, false);
+            editor.commit();
+
+            android.support.v7.app.AlertDialog.Builder builder = new android.support.v7.app.AlertDialog.Builder(this);
+            builder.setTitle(R.string.room_e2e_alert_title);
+            builder.setMessage(R.string.room_e2e_alert_message);
+            builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int which) {
+                    // NOP
+                }
+            });
+            builder.create().show();
         }
     }
 }
